requirements.txt
# Core dependencies
streamlit>=1.32.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
plotly>=5.18.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
scipy>=1.11.0
einops>=0.7.0
albumentations>=1.3.1
staintools>=2.1.2
wandb>=0.16.0
tqdm>=4.66.0
pytest>=7.4.0
pytest-cov>=4.1.0
mypy>=1.7.0
flake8>=6.1.0
black>=23.12.0
tensorboard>=2.14.0
seaborn>=0.13.0
jupyter>=1.0.0
ipywidgets>=8.1.0
gradio>=4.7.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
psutil>=5.9.0

# State-of-the-art model dependencies
timm>=0.9.12  # For Swin Transformer V2, ConvNeXt V2, EfficientNet V2

# Medical imaging
openslide-python>=1.3.0  # WSI support

# Additional utilities
pyyaml>=6.0
optuna>=3.4.0  # Hyperparameter optimization

# Distributed training and WSI processing enhancements
ray[default]>=2.8.0  # Distributed WSI processing
dask[complete]>=2023.12.0  # Alternative distributed computing
lmdb>=1.4.1  # Fast dataset access
h5py>=3.10.0  # HDF5 datasets
zarr>=2.16.0  # Cloud-native array storage
fsspec>=2023.12.0  # Unified filesystem interface
s3fs>=2023.12.0  # S3 filesystem support
gcsfs>=2023.12.0  # Google Cloud Storage support
adlfs>=2023.12.0  # Azure Data Lake storage
deepspeed>=0.12.0  # Advanced distributed training
fairscale>=0.4.13  # Model parallelism utilities
flash-attn>=2.4.2  # Flash attention for memory efficiency
apex  # NVIDIA mixed precision utilities
nvidia-ml-py3>=7.352.0  # GPU monitoring
py3nvml>=0.2.7  # Alternative GPU monitoring
accelerate>=0.25.0  # Hugging Face accelerate for distributed training
datasets>=2.15.0  # Dataset management
evaluate>=0.4.1  # Model evaluation metrics

# Note: PyRadiomics dependencies moved to requirements_local.txt for cloud compatibility 